emotion_assistant/requirements.txt
absl-py==2.3.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
astunparse==1.6.3
async-timeout==5.0.1
beautifulsoup4==4.14.3
billiard==4.2.4
blinker==1.9.0
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
channels==4.3.2
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.2
cryptography==46.0.3
cycler==0.12.1
decorator==4.4.2
deepface==0.0.96
distro==1.9.0
Django==5.2.10
django-allauth==65.13.1
djangorestframework==3.16.1
dotenv==0.9.9
exceptiongroup==1.3.1
facenet-pytorch==2.5.3
fer==25.10.3
ffmpeg-python==0.2.0
filelock==3.20.3
fire==0.7.1
Flask==3.1.2
flask-cors==6.0.2
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2026.1.0
future==1.0.0
gast==0.7.0
gdown==5.2.1
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.187.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
google-pasta==0.2.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
h5py==3.15.1
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
keras==3.12.0
kiwisolver==1.4.9
kombu==5.6.2
libclang==18.1.1
lightecc==0.0.4
lightphe==0.0.20
lz4==4.4.5
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
moviepy==1.0.3
mpmath==1.3.0
mtcnn==1.0.0
namex==0.1.0
networkx==3.4.2
numpy==2.2.6
openai==2.15.0
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pandas==2.3.3
pillow==12.1.0
proglog==0.1.12
prompt_toolkit==3.0.52
proto-plus==1.27.0
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.3.1
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
redis==7.1.0
requests==2.32.5
retina-face==0.0.17
rich==14.2.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.1
sqlparse==0.5.5
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
tf_keras==2.20.1
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.2.14
Werkzeug==3.1.5
wrapt==2.0.1
